refactor: extract interepOnboarding views in separate files
